Refactor directories of e2e-tests
