+++ e2e/annotation-requests-displayed.spec.ts
@@ -2,9 +2,9 @@
 import express from 'express';
 import http from 'http';
 import { By } from 'selenium-webdriver';
-import { buildBrowser } from './browser';
-import * as e2ePPSettings from './settings';
-import { simulateLogIn } from './common';
+import { buildBrowser } from './setup/browser';
+import * as e2ePPSettings from './setup/settings';
+import { simulateLogIn } from './shared/routines';
 import { newTab, switchToTab } from './utils';
 
 const packageConf = require('../package');
