+++ e2e/annotations-api-called.spec.ts
@@ -4,9 +4,9 @@ import { waitUntil } from './utils';
 import express from 'express';
 import http from 'http';
 import { By } from 'selenium-webdriver';
-import { buildBrowser } from './browser';
-import * as e2ePPSettings from './settings';
-import { simulateLogIn } from './common';
+import { buildBrowser } from './setup/browser';
+import * as e2ePPSettings from './setup/settings';
+import { simulateLogIn } from './shared/routines';
 
 describe('extension calls api on startup', () => {
   let browser;
