+++ e2e/auth-tokens.spec.ts
@@ -5,9 +5,9 @@ const packageConf = require('../package');
 import express from 'express';
 import http from 'http';
 import { By } from 'selenium-webdriver';
-import { buildBrowser } from './browser';
-import * as e2ePPSettings from './settings';
-import { simulateLogIn } from './common';
+import { buildBrowser } from './setup/browser';
+import * as e2ePPSettings from './setup/settings';
+import { simulateLogIn } from './shared/routines';
 
 let refreshTimesCalled;
 let refreshStatus;
